TodoappDRF/requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.11.0
django-mssql-backend==2.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Markdown==3.3.6
mssql-django==1.1.1
PyJWT==2.3.0
pyodbc==4.0.32
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
